Fix parsing of Lint error message on windows (#34)
* Fix parsing of Lint error message on windows

On windows the filePath includes `C:` which breaks the current `split(":")` logic which depends on the error message being in the second element of `splittedline`.  This simply removes the filename from `line` before the call to `split`, and then assumes that the error message is in the first element of `splittedline`.  This seems to work on windows.  I have not tested on mac or linux.

* should have being in previous commit