fix: move to Julia 1.1 support
Moves the codebase to supporting Julia v1.1 (and up?). This requires 
updating how Lint.jl is installed in CI and the server to a method that 
is incompatible with the older versions. Currently since Lint.jl hasn't 
been updated in the Julia registry we need to install it from their 
`master` branch.

Also note that the specs have been changed since it seems that Lint.jl 
now reports a different line number for the error in question... that is 
even further off from what it should be.

BREAKING CHANGE: Drop support for Julia v0.6.